object_tf_broadcaster modified to pass cpplint test
